fix(swap): run foundry formatter

+++ src/swap/interfaces/ISwapAdapter.sol
@@ -2,12 +2,7 @@
 pragma solidity ^0.8.30;
 
 interface ISwapAdapter {
-  function swap(
-    address tokenIn,
-    address tokenOut,
-    uint256 amountIn,
-    uint256 minAmountOut,
-    address from,
-    address to
-  ) external returns (uint amountOut);
+    function swap(address tokenIn, address tokenOut, uint256 amountIn, uint256 minAmountOut, address from, address to)
+        external
+        returns (uint256 amountOut);
 }
